requirements.txt
gitpython>=3.1.0
colorama>=0.4.6
click>=8.1.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0